+++ src/lib/sorters.mjs
@@ -1,6 +1,6 @@
 import * as fsPath from 'node:path'
 
-const breadthFirstSorter = (a,b) => {
+const breadthFirstSorter = (a, b) => {
   const aDir = a.isDirectory() ? fsPath.join(a.path, a.name) : a.path
   const bDir = b.isDirectory() ? fsPath.join(b.path, b.name) : a.path
 
@@ -8,7 +8,7 @@ const breadthFirstSorter = (a,b) => {
   return dirComp !== 0 ? dirComp : a.name.localeCompare(b.name)
 }
 
-const depthFirstSorter = (a,b) => {
+const depthFirstSorter = (a, b) => {
   if (a.depth < b.depth) {
     return -1
   }
